requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Uploads==0.2.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==2.2.3
SQLAlchemy==2.0.23
email-validator==2.1.0.post1
Pillow==10.0.0
python-magic==0.4.27
eventlet==0.33.3
beautifulsoup4==4.12.2
requests==2.31.0
click==8.1.7
alembic==1.12.0
python-dateutil==2.8.2
lxml==4.9.3
psycopg2-binary==2.9.9
google-generativeai==0.3.2
pytest
pytest-cov
Flask-Testing